requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
pyodbc==5.0.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
eventlet==0.30.2
dnspython==1.16.0
gunicorn==21.2.0
azure-storage-blob==12.8.1
azure-core==1.29.5
azure-identity==1.15.0
simple-websocket==1.1.0 